Initial Backend Code Version-1

+++ src/main/java/com/Auto/Billing/System/repository/BuildingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BuildingRepository extends JpaRepository<Building, Integer> {
-    public Building findById(int bid);
+public interface BuildingRepository extends JpaRepository<Building, Long> {
+    public Building findById(long bid);
 }
