+++ src/main/java/com/Auto/Billing/System/repository/TenantRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-
-public interface TenantRepository extends JpaRepository<Tenant, Integer> {
-    Tenant findById(int tid);
+public interface TenantRepository extends JpaRepository<Tenant, Long> {
+    Tenant findById(long tid);
 }
